rename endpoints to routes
