+++ app/api/routes/detectors.py
@@ -3,7 +3,7 @@ import logging
 from fastapi import APIRouter, Depends
 from model import Detector
 
-from app.core.utils import get_groundlight_sdk_instance
+from app.core.utils import get_groundlight_sdk_instance, safe_call_api
 from app.schemas.schemas import DetectorCreate
 
 logger = logging.getLogger(__name__)
@@ -14,7 +14,8 @@ router = APIRouter()
 
 @router.post("", response_model=Detector)
 async def create_detector(props: DetectorCreate, gl: Depends = Depends(get_groundlight_sdk_instance)):
-    return gl.create_detector(
+    return safe_call_api(
+        gl.create_detector,
         name=props.name,
         query=props.query,
         confidence_threshold=props.confidence_threshold,
@@ -24,4 +25,4 @@ async def create_detector(props: DetectorCreate, gl: Depends = Depends(get_groun
 
 @router.get("/{id}", response_model=Detector)
 async def get_detector(id: str, gl: Depends = Depends(get_groundlight_sdk_instance)):
-    return gl.get_detector(id=id)
+    return safe_call_api(gl.get_detector, id=id)
