+++ app/main.py
@@ -1,3 +1,6 @@
+import logging
+import os
+
 from fastapi import FastAPI
 from groundlight import Groundlight
 
@@ -5,6 +8,11 @@ from app.api.api import api_router, ping_router
 from app.api.naming import API_BASE_PATH
 
 from .core.motion_detection import AsyncMotionDetector, MotdetParameterSettings
+
+LOG_LEVEL = os.environ.get("LOG_LEVEL", "INFO").upper()
+
+logging.basicConfig(level=LOG_LEVEL)
+
 
 app = FastAPI()
 app.include_router(router=api_router, prefix=API_BASE_PATH)
